requirements.txt
numpy==2.3.5
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
PyYAML==6.0.3
ultralytics==8.3.233
